Split codegen tests into file
